+++ src/neuroconv/tools/nwb_helpers/_models/_base_models.py
@@ -62,7 +62,7 @@ class DatasetInfo(BaseModel):
         super().__init__(**values)
 
 
-class DatasetConfiguration(BaseModel, ABC):
+class DatasetIOConfiguration(BaseModel, ABC):
     """A data model for configuring options about an object that will become a HDF5 or Zarr Dataset in the file."""
 
     # TODO: When using Pydantic v2, remove
@@ -188,7 +188,7 @@ class BackendConfiguration(BaseModel):
 
     backend: Literal["hdf5", "zarr"] = Field(description="The name of the backend used to configure the NWBFile.")
     data_io_class: Type[DataIO] = Field(description="The DataIO class that is specific to this backend.")
-    dataset_configurations: Dict[str, DatasetConfiguration] = Field(
+    dataset_configurations: Dict[str, DatasetIOConfiguration] = Field(
         description=(
             "A mapping from object locations (e.g. `acquisition/TestElectricalSeriesAP/data`) "
             "to their DatasetConfiguration specification that contains all information "